feat: improve UTF string handling and size_t compatibility
* Enhance UTF string conversion and buffer handling
  - Improve buffer conversion efficiency by batching character adds
  - Add mulle_utf8data introspection capabilities
  - Optimize UTF32 to UTF8 conversion paths

* Strengthen type safety and consistency
  - Replace unsigned int with size_t for string lengths
  - Change mulle_utf8_t to char for better C compatibility
  - Update function signatures to use size_t consistently

* Breaking changes
  - Rename mulle_unicode_is_privatecharacterplane to mulle_utf_is_privatecharacterplane
  - Change parameter order in _strcnpy functions
  - Update mulle_utf_add_bytes_function_t typedef
  - Move stdlib-like code to mulle-utf8-string.c

* Other improvements
  - Increase test coverage significantly
  - Fix 32-bit compatibility issues
  - Update CMake configuration for better Windows support

+++ test/iteration/iteration.c
@@ -61,7 +61,7 @@ static void   test( mulle_utf32_t text[ 4])
    mulle_utf16_t   *utf16;
    char    *utf8;
    mulle_utf32_t   c;
-   unsigned int    i;
+   size_t    i;
 
    memset( &buffer16, 0, sizeof( buffer16));
    memset( &buffer8, 0, sizeof( buffer8));
@@ -127,7 +127,7 @@ static void   test( mulle_utf32_t text[ 4])
 static void   stress_test()
 {
    mulle_utf32_t   text[ 4];
-   unsigned int    i;
+   size_t    i;
    mulle_utf32_t   mask;
 
    mask = 0;
